requirements.txt
Django==5.2.6
oracledb==3.3.0
python-decouple
Pillow
django-cleanup
whitenoise